Tracking last successful data synchronization time
Included last successful data synchronization time in the heartbeat messages

+++ internal/datatransfer/s3/sync.go
@@ -21,7 +21,7 @@ type Sync struct {
 	url       string
 }
 
-func NewSync(host string, port int64, accessKey, secretKey, bucket string) *Sync {
+func NewSync(host string, port int32, accessKey, secretKey, bucket string) *Sync {
 	return &Sync{
 		accessKey: accessKey,
 		secretKey: secretKey,
